refactor(benchmark): organize editor as a page module - Separate the route entry from the editor component - Move the existing page test alongside its entry
